Initial work on separating core functionality from distribution dependent code
